requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pg8000
cloud-sql-python-connector[pg8000]
psycopg2-binary
python-dotenv
jinja2
pydantic-settings
asyncpg